cleanup interfaces
